# Conflicts:
#	scipy/stats/tests/test_continuous_basic.py
#	scipy/stats/tests/test_distributions.py

+++ scipy/optimize/rectangular_lsap/rectangular_lsap.h
@@ -37,7 +37,7 @@ extern "C" {
 
 #include <stdint.h>
 
-int solve_rectangular_linear_sum_assignment(int nr, int nc, double* input_cost,
+int solve_rectangular_linear_sum_assignment(intptr_t nr, intptr_t nc, double* input_cost,
                                             int64_t* col4row);
 
 #ifdef __cplusplus
